moved function out of keycloak.js, added catch to client component, api spec update
